server/requirements.txt
torch
werkzeug
Pillow
numpy
cbor2
flask
pyopenssl
flask_cors
umap-learn
pynndescent
